Revert "Fix botattr validation and correct flag usage."
This reverts commit 8f8469fb77c9437836f409659774293f97ac8b22.

+++ src/users.h
@@ -97,7 +97,7 @@ struct user_entry {
     char *string;
     void *extra;
     struct list_type *list;
-    unsigned long long ulong;
+    unsigned long ulong;
   } u;
   char *name;
 };
